py-toolkit/requirements.txt
et-xmlfile==1.1.0
lxml==4.9.2
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
pytest==7.4.0